+++ src/Übungen/Funktionale_Programmierung/root_test.java
@@ -2,13 +2,13 @@ package Übungen.Funktionale_Programmierung;
 
 public class root_test
 {
-    public static void main(String[]args)
-{
-        System.out.println(isRoot(2,9));
-    }
+	public static void main(String[] args)
+	{
+		System.out.println(isRoot(2, 9));
+	}
 
-    public static boolean isRoot(int root, long num)
-{
-        return root > 0 && root * root == num;
-    }
+	public static boolean isRoot(int root, long num)
+	{
+		return root > 0 && (long) root * root == num;
+	}
 }
